+++ src/Repository/PostRepository.php
@@ -13,6 +13,7 @@ use Doctrine\Persistence\ManagerRegistry;
  */
 class PostRepository extends ServiceEntityRepository
 {
+    use SearchQueryTrait;
     public function __construct(ManagerRegistry $registry)
     {
         parent::__construct($registry, Post::class);
@@ -54,25 +55,6 @@ class PostRepository extends ServiceEntityRepository
     public function searchByQuery(string $query, int $offset, int $limit): array
     {
         return $this->buildSearchQuery('caption', $query, $offset, $limit, 'createdAt', 'DESC')->getResult();
-    }
-
-    /**
-     * @param string $searchField Field to search on
-     * @param string $query Search query term
-     * @param int $offset Pagination offset
-     * @param int $limit Pagination limit
-     * @param string $orderField Field to order by
-     * @param string $direction Sort direction
-     */
-    protected function buildSearchQuery(string $searchField, string $query, int $offset, int $limit, string $orderField, string $direction): \Doctrine\ORM\Query
-    {
-        return $this->createQueryBuilder('e')
-            ->where('e.'.$searchField.' LIKE :query')
-            ->setParameter('query', '%'.$query.'%')
-            ->orderBy('e.'.$orderField, $direction)
-            ->setFirstResult($offset)
-            ->setMaxResults($limit)
-            ->getQuery();
     }
 
     public function updateViewsCount(int $postId, int $count): void
