+++ src/Repository/UserRepository.php
@@ -16,6 +16,7 @@ use Symfony\Component\Security\Core\User\PasswordUpgraderInterface;
  */
 class UserRepository extends ServiceEntityRepository implements PasswordUpgraderInterface
 {
+    use SearchQueryTrait;
     public function __construct(ManagerRegistry $registry)
     {
         parent::__construct($registry, User::class);
@@ -53,25 +54,6 @@ class UserRepository extends ServiceEntityRepository implements PasswordUpgrader
     public function searchByQuery(string $query, int $offset, int $limit): array
     {
         return $this->buildSearchQuery('username', $query, $offset, $limit, 'username', 'ASC')->getResult();
-    }
-
-    /**
-     * @param string $searchField Field to search on
-     * @param string $query Search query term
-     * @param int $offset Pagination offset
-     * @param int $limit Pagination limit
-     * @param string $orderField Field to order by
-     * @param string $direction Sort direction
-     */
-    protected function buildSearchQuery(string $searchField, string $query, int $offset, int $limit, string $orderField, string $direction): \Doctrine\ORM\Query
-    {
-        return $this->createQueryBuilder('e')
-            ->where('e.'.$searchField.' LIKE :query')
-            ->setParameter('query', '%'.$query.'%')
-            ->orderBy('e.'.$orderField, $direction)
-            ->setFirstResult($offset)
-            ->setMaxResults($limit)
-            ->getQuery();
     }
 
     public function deleteTokenByUserToken(string $deviceToken): void
